# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: Sistema de Gestión de Seguros - API Centralizada
  description: |
    Especificación consolidada para la gestión de Clientes y Pólizas.
    Incluye trazabilidad mediante el campo `transactionId` y respuestas estandarizadas.
  version: 1.0.0
  contact:
    name: Anthony Colmenares
    email: acolmenares@insurance.com

servers:
  - url: http://localhost:8081
    description: Customer Service
  - url: http://localhost:8082
    description: Policy Service

components:
  schemas:
    ApiResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        message:
          type: string
          example: "success"
        code:
          type: integer
          example: 200
        data:
          type: object

    Customer:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        birthDate:
          type: string
          format: date

    Policy:
      type: object
      required:
        - customerId
        - type
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        type:
          type: string
          enum: [VIDA, VEHICULO, SALUD]
        beneficiaries:
          type: array
          items:
            $ref: '#/components/schemas/Person'
        insuredVehicles:
          type: array
          items:
            type: string
        extraParents:
          type: integer
        extraChildren:
          type: integer
        hasSpouse:
          type: boolean

    Person:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dni:
          type: string

paths:
  /api/customers:
    post:
      summary: Crear un nuevo cliente
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Cliente creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      summary: Obtener todos los clientes
      tags: [Customers]
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/policies:
    post:
      summary: Crear una nueva póliza
      tags: [Policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        '201':
          description: Póliza creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/policies/customer/{customerId}:
    get:
      summary: Obtener pólizas de un cliente
      tags: [Policies]
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista de pólizas del cliente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'