# services/policy-service/src/main/java/com/insurance/policy/domain/entity/Person.java
package com.insurance.policy.domain.entity;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Person {
    private String firstName;
    private String lastName;
    private String documentNumber;
}